+++ src/plugin/systems.rs
@@ -7,29 +7,30 @@ use crate::dynamics::{
     ReadMassProperties, RigidBody, Sleeping, TransformInterpolation, Velocity,
 };
 use crate::geometry::{
-    ActiveCollisionTypes, ActiveEvents, ActiveHooks, Collider, ColliderMassProperties,
-    ColliderScale, CollisionGroups, ContactForceEventThreshold, Friction, RapierColliderHandle,
-    Restitution, Sensor, SolverGroups,
+    ActiveCollisionTypes, ActiveEvents, ActiveHooks, Collider, ColliderDisabled,
+    ColliderMassProperties, ColliderScale, CollisionGroups, ContactForceEventThreshold, Friction,
+    RapierColliderHandle, Restitution, Sensor, SolverGroups,
 };
-use crate::pipeline::{
-    CollisionEvent, ContactForceEvent, PhysicsHooksWithQueryInstance, PhysicsHooksWithQueryResource,
-};
+use crate::pipeline::{CollisionEvent, ContactForceEvent};
 use crate::plugin::configuration::{SimulationToRenderTime, TimestepMode};
 use crate::plugin::{RapierConfiguration, RapierContext};
 use crate::prelude::{
-    CollidingEntities, KinematicCharacterController, KinematicCharacterControllerOutput,
+    BevyPhysicsHooks, BevyPhysicsHooksAdapter, CollidingEntities, KinematicCharacterController,
+    KinematicCharacterControllerOutput, RigidBodyDisabled,
 };
 use crate::utils;
-use bevy::ecs::query::WorldQuery;
+use bevy::ecs::system::{StaticSystemParam, SystemParamItem};
 use bevy::prelude::*;
 use rapier::prelude::*;
 use std::collections::HashMap;
 
-#[cfg(feature = "dim3")]
-use crate::prelude::{AsyncCollider, AsyncSceneCollider};
+#[cfg(all(feature = "dim3", feature = "async-collider"))]
+use {
+    crate::prelude::{AsyncCollider, AsyncSceneCollider},
+    bevy::scene::SceneInstance,
+};
 
 use crate::control::CharacterCollision;
-use crate::utils::transform_to_iso;
 #[cfg(feature = "dim2")]
 use bevy::math::Vec3Swizzles;
 
@@ -58,6 +59,7 @@ pub type RigidBodyComponents<'a> = (
     Option<&'a Dominance>,
     Option<&'a Sleeping>,
     Option<&'a Damping>,
+    Option<&'a RigidBodyDisabled>,
 );
 
 /// Components related to colliders.
@@ -74,6 +76,7 @@ pub type ColliderComponents<'a> = (
     Option<&'a CollisionGroups>,
     Option<&'a SolverGroups>,
     Option<&'a ContactForceEventThreshold>,
+    Option<&'a ColliderDisabled>,
 );
 
 /// System responsible for applying [`GlobalTransform::scale`] and/or [`ColliderScale`] to
@@ -108,7 +111,7 @@ pub fn apply_scale(
             None => transform.compute_transform().scale,
         };
 
-        if shape.scale != effective_scale {
+        if shape.scale != crate::geometry::get_snapped_scale(effective_scale) {
             shape.set_scale(effective_scale, config.scaled_shape_subdivision);
         }
     }
@@ -137,6 +140,7 @@ pub fn apply_collider_user_changes(
     >,
     changed_solver_groups: Query<(&RapierColliderHandle, &SolverGroups), Changed<SolverGroups>>,
     changed_sensors: Query<(&RapierColliderHandle, &Sensor), Changed<Sensor>>,
+    changed_disabled: Query<(&RapierColliderHandle, &ColliderDisabled), Changed<ColliderDisabled>>,
     changed_contact_force_threshold: Query<
         (&RapierColliderHandle, &ContactForceEventThreshold),
         Changed<ContactForceEventThreshold>,
@@ -217,6 +221,12 @@ pub fn apply_collider_user_changes(
         }
     }
 
+    for (handle, _) in changed_disabled.iter() {
+        if let Some(co) = context.colliders.get_mut(handle.0) {
+            co.set_enabled(false);
+        }
+    }
+
     for (handle, threshold) in changed_contact_force_threshold.iter() {
         if let Some(co) = context.colliders.get_mut(handle.0) {
             co.set_contact_force_event_threshold(threshold.0);
@@ -265,6 +275,10 @@ pub fn apply_rigid_body_user_changes(
     changed_dominance: Query<(&RapierRigidBodyHandle, &Dominance), Changed<Dominance>>,
     changed_sleeping: Query<(&RapierRigidBodyHandle, &Sleeping), Changed<Sleeping>>,
     changed_damping: Query<(&RapierRigidBodyHandle, &Damping), Changed<Damping>>,
+    changed_disabled: Query<
+        (&RapierRigidBodyHandle, &RigidBodyDisabled),
+        Changed<RigidBodyDisabled>,
+    >,
 ) {
     let context = &mut *context;
     let scale = context.physics_scale;
@@ -293,7 +307,7 @@ pub fn apply_rigid_body_user_changes(
     //       position instead of the current one.
     for (handle, rb_type) in changed_rb_types.iter() {
         if let Some(rb) = context.bodies.get_mut(handle.0) {
-            rb.set_body_type((*rb_type).into());
+            rb.set_body_type((*rb_type).into(), true);
         }
     }
 
@@ -428,6 +442,12 @@ pub fn apply_rigid_body_user_changes(
             rb.set_angular_damping(damping.angular_damping);
         }
     }
+
+    for (handle, _) in changed_disabled.iter() {
+        if let Some(co) = context.bodies.get_mut(handle.0) {
+            co.set_enabled(false);
+        }
+    }
 }
 
 /// System responsible for applying changes the user made to a joint component.
@@ -512,10 +532,6 @@ pub fn writeback_rigid_bodies(
                         if let Some(parent_global_transform) =
                             parent.and_then(|p| global_transforms.get(**p).ok())
                         {
-                            // In 2D, preserve the transform `z` component that may have been set by the user
-                            #[cfg(feature = "dim2")]
-                            let prev_z = transform.translation.z;
-
                             // We need to compute the new local transform such that:
                             // curr_parent_global_transform * new_transform = interpolated_pos
                             // new_transform = curr_parent_global_transform.inverse() * interpolated_pos
@@ -525,9 +541,17 @@ pub fn writeback_rigid_bodies(
                                     .inverse()
                                     .to_scale_rotation_translation();
                             let new_rotation = inverse_parent_rotation * interpolated_pos.rotation;
-                            let new_translation = inverse_parent_rotation
+
+                            #[allow(unused_mut)] // mut is needed in 2D but not in 3D.
+                            let mut new_translation = inverse_parent_rotation
                                 * interpolated_pos.translation
                                 + inverse_parent_translation;
+
+                            // In 2D, preserve the transform `z` component that may have been set by the user
+                            #[cfg(feature = "dim2")]
+                            {
+                                new_translation.z = transform.translation.z;
+                            }
 
                             if transform.rotation != new_rotation
                                 || transform.translation != new_translation
@@ -537,11 +561,6 @@ pub fn writeback_rigid_bodies(
                                 //       change tracking when the values didn’t change.
                                 transform.rotation = new_rotation;
                                 transform.translation = new_translation;
-                            }
-
-                            #[cfg(feature = "dim2")]
-                            {
-                                transform.translation.z = prev_z;
                             }
 
                             // NOTE: we need to compute the result of the next transform propagation
@@ -609,29 +628,28 @@ pub fn writeback_rigid_bodies(
 
 /// System responsible for advancing the physics simulation, and updating the internal state
 /// for scene queries.
-pub fn step_simulation<PhysicsHooksData: 'static + WorldQuery + Send + Sync>(
+pub fn step_simulation<Hooks>(
     mut context: ResMut<RapierContext>,
     config: Res<RapierConfiguration>,
-    hooks: Res<PhysicsHooksWithQueryResource<PhysicsHooksData>>,
-    (time, mut sim_to_render_time): (Res<Time>, ResMut<SimulationToRenderTime>),
+    hooks: StaticSystemParam<Hooks>,
+    time: Res<Time>,
+    mut sim_to_render_time: ResMut<SimulationToRenderTime>,
     collision_events: EventWriter<CollisionEvent>,
     contact_force_events: EventWriter<ContactForceEvent>,
-    hooks_data: Query<PhysicsHooksData>,
     interpolation_query: Query<(&RapierRigidBodyHandle, &mut TransformInterpolation)>,
-) {
+) where
+    Hooks: 'static + BevyPhysicsHooks,
+    for<'w, 's> SystemParamItem<'w, 's, Hooks>: BevyPhysicsHooks,
+{
     let context = &mut *context;
+    let hooks_adapter = BevyPhysicsHooksAdapter::new(hooks.into_inner());
 
     if config.physics_pipeline_active {
-        let hooks_instance = PhysicsHooksWithQueryInstance {
-            user_data: hooks_data,
-            hooks: &*hooks.0,
-        };
-
         context.step_simulation(
             config.gravity,
             config.timestep_mode,
             Some((collision_events, contact_force_events)),
-            &hooks_instance,
+            &hooks_adapter,
             &time,
             &mut sim_to_render_time,
             Some(interpolation_query),
@@ -647,20 +665,26 @@ pub fn step_simulation<PhysicsHooksData: 'static + WorldQuery + Send + Sync>(
 }
 
 /// NOTE: This currently does nothing in 2D.
+#[cfg(feature = "async-collider")]
 #[cfg(feature = "dim2")]
+pub fn init_async_colliders() {}
+
+/// NOTE: This does nothing in 3D with headless.
+#[cfg(feature = "headless")]
+#[cfg(feature = "dim3")]
 pub fn init_async_colliders() {}
 
 /// System responsible for creating `Collider` components from `AsyncCollider` components if the
 /// corresponding mesh has become available.
-#[cfg(feature = "dim3")]
+#[cfg(all(feature = "dim3", feature = "async-collider"))]
 pub fn init_async_colliders(
     mut commands: Commands,
     meshes: Res<Assets<Mesh>>,
-    async_colliders: Query<(Entity, &AsyncCollider)>,
+    async_colliders: Query<(Entity, &Handle<Mesh>, &AsyncCollider)>,
 ) {
-    for (entity, async_collider) in async_colliders.iter() {
-        if let Some(mesh) = meshes.get(&async_collider.handle) {
-            match Collider::from_bevy_mesh(mesh, &async_collider.shape) {
+    for (entity, mesh_handle, async_collider) in async_colliders.iter() {
+        if let Some(mesh) = meshes.get(mesh_handle) {
+            match Collider::from_bevy_mesh(mesh, &async_collider.0) {
                 Some(collider) => {
                     commands
                         .entity(entity)
@@ -675,19 +699,19 @@ pub fn init_async_colliders(
 
 /// System responsible for creating `Collider` components from `AsyncSceneCollider` components if the
 /// corresponding scene has become available.
-#[cfg(feature = "dim3")]
+#[cfg(all(feature = "dim3", feature = "async-collider"))]
 pub fn init_async_scene_colliders(
     mut commands: Commands,
     meshes: Res<Assets<Mesh>>,
-    scenes: Res<Assets<Scene>>,
-    async_colliders: Query<(Entity, &AsyncSceneCollider)>,
+    scene_spawner: Res<SceneSpawner>,
+    async_colliders: Query<(Entity, &SceneInstance, &AsyncSceneCollider)>,
     children: Query<&Children>,
     mesh_handles: Query<(&Name, &Handle<Mesh>)>,
 ) {
-    for (entity, async_collider) in async_colliders.iter() {
-        if scenes.get(&async_collider.handle).is_some() {
-            traverse_descendants(entity, &children, &mut |child| {
-                if let Ok((name, handle)) = mesh_handles.get(child) {
+    for (scene_entity, scene_instance, async_collider) in async_colliders.iter() {
+        if scene_spawner.instance_is_ready(**scene_instance) {
+            for child_entity in children.iter_descendants(scene_entity) {
+                if let Ok((name, handle)) = mesh_handles.get(child_entity) {
                     let shape = async_collider
                         .named_shapes
                         .get(name.as_str())
@@ -696,7 +720,7 @@ pub fn init_async_scene_colliders(
                         let mesh = meshes.get(handle).unwrap(); // NOTE: Mesh is already loaded
                         match Collider::from_bevy_mesh(mesh, shape) {
                             Some(collider) => {
-                                commands.entity(child).insert(collider);
+                                commands.entity(child_entity).insert(collider);
                             }
                             None => error!(
                                 "Unable to generate collider from mesh {:?} with name {}",
@@ -705,20 +729,9 @@ pub fn init_async_scene_colliders(
                         }
                     }
                 }
-            });
+            }
 
-            commands.entity(entity).remove::<AsyncSceneCollider>();
-        }
-    }
-}
-
-/// Iterates over all descendants of the `entity` and applies `f`.
-#[cfg(feature = "dim3")]
-fn traverse_descendants(entity: Entity, children: &Query<&Children>, f: &mut impl FnMut(Entity)) {
-    if let Ok(entity_children) = children.get(entity) {
-        for child in entity_children.iter().copied() {
-            f(child);
-            traverse_descendants(child, children, f);
+            commands.entity(scene_entity).remove::<AsyncSceneCollider>();
         }
     }
 }
@@ -728,7 +741,7 @@ pub fn init_colliders(
     mut commands: Commands,
     config: Res<RapierConfiguration>,
     mut context: ResMut<RapierContext>,
-    colliders: Query<ColliderComponents, Without<RapierColliderHandle>>,
+    colliders: Query<(ColliderComponents, Option<&GlobalTransform>), Without<RapierColliderHandle>>,
     mut rigid_body_mprops: Query<&mut ReadMassProperties>,
     parent_query: Query<(&Parent, Option<&Transform>)>,
 ) {
@@ -736,18 +749,22 @@ pub fn init_colliders(
     let physics_scale = context.physics_scale;
 
     for (
-        entity,
-        shape,
-        sensor,
-        mprops,
-        active_events,
-        active_hooks,
-        active_collision_types,
-        friction,
-        restitution,
-        collision_groups,
-        solver_groups,
-        contact_force_event_threshold,
+        (
+            entity,
+            shape,
+            sensor,
+            mprops,
+            active_events,
+            active_hooks,
+            active_collision_types,
+            friction,
+            restitution,
+            collision_groups,
+            solver_groups,
+            contact_force_event_threshold,
+            disabled,
+        ),
+        global_transform,
     ) in colliders.iter()
     {
         let mut scaled_shape = shape.clone();
@@ -755,6 +772,7 @@ pub fn init_colliders(
         let mut builder = ColliderBuilder::new(scaled_shape.raw.clone());
 
         builder = builder.sensor(sensor.is_some());
+        builder = builder.enabled(disabled.is_none());
 
         if let Some(mprops) = mprops {
             builder = match mprops {
@@ -818,10 +836,10 @@ pub fn init_colliders(
             body_handle = context.entity2body.get(&body_entity).copied();
         }
 
-        builder = builder.position(utils::transform_to_iso(&child_transform, physics_scale));
         builder = builder.user_data(entity.to_bits() as u128);
 
         let handle = if let Some(body_handle) = body_handle {
+            builder = builder.position(utils::transform_to_iso(&child_transform, physics_scale));
             let handle =
                 context
                     .colliders
@@ -830,12 +848,19 @@ pub fn init_colliders(
                 // Inserting the collider changed the rigid-body’s mass properties.
                 // Read them back from the engine.
                 if let Some(parent_body) = context.bodies.get(body_handle) {
-                    mprops.0 =
-                        MassProperties::from_rapier(*parent_body.mass_properties(), physics_scale);
+                    mprops.0 = MassProperties::from_rapier(
+                        parent_body.mass_properties().local_mprops,
+                        physics_scale,
+                    );
                 }
             }
             handle
         } else {
+            let global_transform = global_transform.cloned().unwrap_or_default();
+            builder = builder.position(utils::transform_to_iso(
+                &global_transform.compute_transform(),
+                physics_scale,
+            ));
             context.colliders.insert(builder)
         };
 
@@ -866,9 +891,12 @@ pub fn init_rigid_bodies(
         dominance,
         sleep,
         damping,
+        disabled,
     ) in rigid_bodies.iter()
     {
         let mut builder = RigidBodyBuilder::new((*rb).into());
+        builder = builder.enabled(disabled.is_none());
+
         if let Some(transform) = transform {
             builder = builder.position(utils::transform_to_iso(
                 &transform.compute_transform(),
@@ -1046,10 +1074,10 @@ pub fn init_joints(
 pub fn sync_removals(
     mut commands: Commands,
     mut context: ResMut<RapierContext>,
-    removed_bodies: RemovedComponents<RapierRigidBodyHandle>,
-    removed_colliders: RemovedComponents<RapierColliderHandle>,
-    removed_impulse_joints: RemovedComponents<RapierImpulseJointHandle>,
-    removed_multibody_joints: RemovedComponents<RapierMultibodyJointHandle>,
+    mut removed_bodies: RemovedComponents<RapierRigidBodyHandle>,
+    mut removed_colliders: RemovedComponents<RapierColliderHandle>,
+    mut removed_impulse_joints: RemovedComponents<RapierImpulseJointHandle>,
+    mut removed_multibody_joints: RemovedComponents<RapierMultibodyJointHandle>,
     orphan_bodies: Query<Entity, (With<RapierRigidBodyHandle>, Without<RigidBody>)>,
     orphan_colliders: Query<Entity, (With<RapierColliderHandle>, Without<Collider>)>,
     orphan_impulse_joints: Query<Entity, (With<RapierImpulseJointHandle>, Without<ImpulseJoint>)>,
@@ -1058,7 +1086,9 @@ pub fn sync_removals(
         (With<RapierMultibodyJointHandle>, Without<MultibodyJoint>),
     >,
 
-    removed_sensors: RemovedComponents<Sensor>,
+    mut removed_sensors: RemovedComponents<Sensor>,
+    mut removed_rigid_body_disabled: RemovedComponents<RigidBodyDisabled>,
+    mut removed_colliders_disabled: RemovedComponents<ColliderDisabled>,
 ) {
     /*
      * Rigid-bodies removal detection.
@@ -1151,12 +1181,28 @@ pub fn sync_removals(
     }
 
     /*
-     * Sensor marker component removal detection.
+     * Marker components removal detection.
      */
     for entity in removed_sensors.iter() {
         if let Some(handle) = context.entity2collider.get(&entity) {
             if let Some(co) = context.colliders.get_mut(*handle) {
                 co.set_sensor(false);
+            }
+        }
+    }
+
+    for entity in removed_colliders_disabled.iter() {
+        if let Some(handle) = context.entity2collider.get(&entity) {
+            if let Some(co) = context.colliders.get_mut(*handle) {
+                co.set_enabled(true);
+            }
+        }
+    }
+
+    for entity in removed_rigid_body_disabled.iter() {
+        if let Some(handle) = context.entity2body.get(&entity) {
+            if let Some(rb) = context.bodies.get_mut(*handle) {
+                rb.set_enabled(true);
             }
         }
     }
@@ -1250,7 +1296,8 @@ pub fn update_character_controls(
                     shape_pos = body.position() * shape_pos
                 } else if let Some(gtransform) = glob_transform {
                     shape_pos =
-                        transform_to_iso(&gtransform.compute_transform(), physics_scale) * shape_pos
+                        utils::transform_to_iso(&gtransform.compute_transform(), physics_scale)
+                            * shape_pos
                 }
 
                 (&*scaled_shape.raw, shape_pos)
@@ -1274,7 +1321,7 @@ pub fn update_character_controls(
 
             let mut filter = QueryFilter {
                 flags: controller.filter_flags,
-                groups: controller.filter_groups,
+                groups: controller.filter_groups.map(|g| g.into()),
                 exclude_collider: None,
                 exclude_rigid_body: None,
                 predicate: None,
@@ -1355,11 +1402,10 @@ pub fn update_character_controls(
 
 #[cfg(test)]
 mod tests {
-    #[cfg(feature = "dim3")]
+    #[cfg(all(feature = "dim3", feature = "async-collider"))]
     use bevy::prelude::shape::{Capsule, Cube};
     use bevy::{
         asset::AssetPlugin,
-        core::CorePlugin,
         ecs::event::Events,
         render::{settings::WgpuSettings, RenderPlugin},
         scene::ScenePlugin,
@@ -1370,8 +1416,6 @@ mod tests {
 
     use super::*;
     use crate::plugin::{NoUserData, RapierPhysicsPlugin};
-    #[cfg(feature = "dim3")]
-    use crate::prelude::ComputedColliderShape;
 
     #[test]
     fn colliding_entities_updates() {
@@ -1460,7 +1504,7 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "dim3")]
+    #[cfg(all(feature = "dim3", feature = "async-collider"))]
     fn async_collider_initializes() {
         let mut app = App::new();
         app.add_plugin(HeadlessRenderPlugin)
@@ -1469,13 +1513,7 @@ mod tests {
         let mut meshes = app.world.resource_mut::<Assets<Mesh>>();
         let cube = meshes.add(Cube::default().into());
 
-        let entity = app
-            .world
-            .spawn(AsyncCollider {
-                handle: cube,
-                shape: ComputedColliderShape::TriMesh,
-            })
-            .id();
+        let entity = app.world.spawn((cube, AsyncCollider::default())).id();
 
         app.update();
 
@@ -1491,11 +1529,13 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "dim3")]
+    #[cfg(all(feature = "dim3", feature = "async-collider"))]
     fn async_scene_collider_initializes() {
+        use bevy::scene::scene_spawner_system;
+
         let mut app = App::new();
         app.add_plugin(HeadlessRenderPlugin)
-            .add_system(init_async_scene_colliders);
+            .add_system(init_async_scene_colliders.after(scene_spawner_system));
 
         let mut meshes = app.world.resource_mut::<Assets<Mesh>>();
         let cube_handle = meshes.add(Cube::default().into());
@@ -1510,11 +1550,13 @@ mod tests {
         named_shapes.insert("Capsule".to_string(), None);
         let parent = app
             .world
-            .spawn(AsyncSceneCollider {
-                handle: scene,
-                shape: Some(ComputedColliderShape::TriMesh),
-                named_shapes,
-            })
+            .spawn((
+                scene,
+                AsyncSceneCollider {
+                    named_shapes,
+                    ..Default::default()
+                },
+            ))
             .push_children(&[cube, capsule])
             .id();
 
@@ -1672,16 +1714,16 @@ mod tests {
 
     impl Plugin for HeadlessRenderPlugin {
         fn build(&self, app: &mut App) {
-            app.insert_resource(WgpuSettings {
-                backends: None,
-                ..WgpuSettings::default()
-            })
-            .add_plugin(CorePlugin::default())
-            .add_plugin(WindowPlugin::default())
-            .add_plugin(AssetPlugin::default())
-            .add_plugin(ScenePlugin::default())
-            .add_plugin(RenderPlugin::default())
-            .add_plugin(ImagePlugin::default());
+            app.add_plugin(WindowPlugin::default())
+                .add_plugin(AssetPlugin::default())
+                .add_plugin(ScenePlugin::default())
+                .add_plugin(RenderPlugin {
+                    wgpu_settings: WgpuSettings {
+                        backends: None,
+                        ..Default::default()
+                    },
+                })
+                .add_plugin(ImagePlugin::default());
         }
     }
 }