+++ src/render/lines/mod.rs
@@ -15,7 +15,7 @@
  *
  */
 use bevy::{
-    asset::{Assets, HandleUntyped},
+    asset::{load_internal_asset, Assets, HandleUntyped},
     prelude::*,
     reflect::TypeUuid,
     render::{
@@ -34,35 +34,15 @@ mod render_dim;
 // This module exists to "isolate" the `#[cfg]` attributes to this part of the
 // code. Otherwise, we would pollute the code with a lot of feature
 // gates-specific code.
-#[cfg(feature = "dim3")]
-mod dim {
+#[cfg(feature = "debug-render-3d")]
+mod dim3 {
     pub(crate) use super::render_dim::r3d::{queue, DebugLinePipeline, DrawDebugLines};
     pub(crate) use bevy::core_pipeline::core_3d::Opaque3d as Phase;
-    use bevy::{asset::Handle, render::mesh::Mesh};
-
-    pub(crate) type MeshHandle = Handle<Mesh>;
-    pub(crate) fn from_handle(from: &MeshHandle) -> &Handle<Mesh> {
-        from
-    }
-    pub(crate) fn into_handle(from: Handle<Mesh>) -> MeshHandle {
-        from
-    }
-    pub(crate) const DIMMENSION: &str = "3d";
 }
-#[cfg(feature = "dim2")]
-mod dim {
+#[cfg(feature = "debug-render-2d")]
+mod dim2 {
     pub(crate) use super::render_dim::r2d::{queue, DebugLinePipeline, DrawDebugLines};
     pub(crate) use bevy::core_pipeline::core_2d::Transparent2d as Phase;
-    use bevy::{asset::Handle, render::mesh::Mesh, sprite::Mesh2dHandle};
-
-    pub(crate) type MeshHandle = Mesh2dHandle;
-    pub(crate) fn from_handle(from: &MeshHandle) -> &Handle<Mesh> {
-        &from.0
-    }
-    pub(crate) fn into_handle(from: Handle<Mesh>) -> MeshHandle {
-        Mesh2dHandle(from)
-    }
-    pub(crate) const DIMMENSION: &str = "2d";
 }
 
 pub(crate) const SHADER_FILE: &str = include_str!("debuglines.wgsl");
@@ -125,25 +105,47 @@ impl DebugLinesPlugin {
     }
 }
 
+#[derive(SystemSet, Hash, Debug, PartialEq, Eq, Clone)]
+pub struct DrawLinesLabel;
+
 impl Plugin for DebugLinesPlugin {
     fn build(&self, app: &mut App) {
-        use bevy::render::{render_resource::SpecializedMeshPipelines, RenderApp, RenderStage};
+        use bevy::render::{render_resource::SpecializedMeshPipelines, RenderApp, RenderSet};
         let mut shaders = app.world.get_resource_mut::<Assets<Shader>>().unwrap();
         shaders.set_untracked(DEBUG_LINES_SHADER_HANDLE, Shader::from_wgsl(SHADER_FILE));
-        let lines_config = DebugLinesConfig::always_on_top(self.always_on_top);
-        app.init_resource::<DebugLines>();
-        app.add_startup_system(setup)
-            .add_system_to_stage(CoreStage::PostUpdate, update.label("draw_lines"))
-            .insert_resource(lines_config.clone());
-        app.sub_app_mut(RenderApp)
-            .add_render_command::<dim::Phase, dim::DrawDebugLines>()
-            .insert_resource(lines_config)
-            .init_resource::<dim::DebugLinePipeline>()
-            .init_resource::<SpecializedMeshPipelines<dim::DebugLinePipeline>>()
-            .add_system_to_stage(RenderStage::Extract, extract)
-            .add_system_to_stage(RenderStage::Queue, dim::queue);
 
-        info!("Loaded {} debug lines plugin.", dim::DIMMENSION);
+        let lines_config = DebugLinesConfig::always_on_top(self.always_on_top);
+        app.init_resource::<DebugLines>()
+            .add_startup_system(setup)
+            .add_system(
+                update
+                    .in_base_set(CoreSet::PostUpdate)
+                    .in_set(DrawLinesLabel),
+            )
+            .insert_resource(lines_config.clone());
+
+        #[cfg(feature = "debug-render-3d")]
+        app.sub_app_mut(RenderApp)
+            .add_render_command::<dim3::Phase, dim3::DrawDebugLines>()
+            .init_resource::<dim3::DebugLinePipeline>()
+            .init_resource::<SpecializedMeshPipelines<dim3::DebugLinePipeline>>()
+            .add_system(dim3::queue.in_set(RenderSet::Queue));
+
+        #[cfg(feature = "debug-render-2d")]
+        app.sub_app_mut(RenderApp)
+            .add_render_command::<dim2::Phase, dim2::DrawDebugLines>()
+            .init_resource::<dim2::DebugLinePipeline>()
+            .init_resource::<SpecializedMeshPipelines<dim2::DebugLinePipeline>>()
+            .add_system(dim2::queue.in_set(RenderSet::Queue));
+
+        app.sub_app_mut(RenderApp)
+            .insert_resource(lines_config)
+            .add_system(extract.in_schedule(ExtractSchedule));
+
+        #[cfg(feature = "debug-render-3d")]
+        info!("Loaded 3d debug lines plugin.");
+        #[cfg(feature = "debug-render-2d")]
+        info!("Loaded 2d debug lines plugin.");
     }
 }
 
@@ -174,26 +176,39 @@ fn setup(mut cmds: Commands, mut meshes: ResMut<Assets<Mesh>>) {
         // https://github.com/Toqozz/bevy_debug_lines/issues/16
         //mesh.set_indices(Some(Indices::U16(Vec::with_capacity(MAX_POINTS_PER_MESH))));
 
+        let mesh_handle = meshes.add(mesh);
+
         cmds.spawn((
-            dim::into_handle(meshes.add(mesh)),
-            #[cfg(feature = "dim3")]
-            (bevy::pbr::NotShadowCaster, bevy::pbr::NotShadowReceiver),
-            SpatialBundle::VISIBLE_IDENTITY,
+            SpatialBundle::INHERITED_IDENTITY,
             NoFrustumCulling,
             DebugLinesMesh(i),
+            #[cfg(feature = "debug-render-3d")]
+            (
+                mesh_handle.clone(),
+                bevy::pbr::NotShadowCaster,
+                bevy::pbr::NotShadowReceiver,
+            ),
+            #[cfg(feature = "debug-render-2d")]
+            bevy::sprite::Mesh2dHandle(mesh_handle),
         ));
     }
 }
 
 fn update(
-    debug_line_meshes: Query<(&dim::MeshHandle, &DebugLinesMesh)>,
+    #[cfg(feature = "debug-render-3d")] debug_line_meshes_3d: Query<(
+        &Handle<Mesh>,
+        &DebugLinesMesh,
+    )>,
+    #[cfg(feature = "debug-render-2d")] debug_line_meshes_2d: Query<(
+        &bevy::sprite::Mesh2dHandle,
+        &DebugLinesMesh,
+    )>,
     time: Res<Time>,
     mut meshes: ResMut<Assets<Mesh>>,
     mut lines: ResMut<DebugLines>,
 ) {
     // For each debug line mesh, fill its buffers with the relevant positions/colors chunks.
-    for (mesh_handle, debug_lines_idx) in debug_line_meshes.iter() {
-        let mesh = meshes.get_mut(dim::from_handle(mesh_handle)).unwrap();
+    let update = |mesh: &mut Mesh, debug_lines_idx: &DebugLinesMesh| {
         use VertexAttributeValues::{Float32x3, Float32x4};
         if let Some(Float32x3(vbuffer)) = mesh.attribute_mut(Mesh::ATTRIBUTE_POSITION) {
             vbuffer.clear();
@@ -228,6 +243,18 @@ fn update(
             }
         }
         */
+    };
+
+    #[cfg(feature = "debug-render-3d")]
+    for (mesh_handle, debug_lines_idx) in &debug_line_meshes_3d {
+        let mesh = meshes.get_mut(mesh_handle).unwrap();
+        update(mesh, debug_lines_idx);
+    }
+
+    #[cfg(feature = "debug-render-2d")]
+    for (mesh2d_handle, debug_lines_idx) in &debug_line_meshes_2d {
+        let mesh = meshes.get_mut(&mesh2d_handle.0).unwrap();
+        update(mesh, debug_lines_idx);
     }
 
     // Processes stuff like getting rid of expired lines and stuff.